fix(KEN-791): a value keeps its own newlines [no-changelog]
#1. Continuation lines were stored without their terminators and
render_entries appended the destination file's EOL to every one, so an LF
template seeded into a CRLF file rewrote the newlines INSIDE the string.
The consumer's value was not the template's value, which is the guarantee
this PR is built on.

Two kinds of newline meet in a seeded block and only one is the file's.
The terminators between entries and around the block belong to the
destination and still take its spelling; the ones inside a value are the
value's own content. `EnvEntry::assignment` is now one String holding the
assignment as written, terminators included and none after the last line,
built from the walk's `raw` rather than its `text`. render_entries writes
the comment lines in the file's EOL and the value byte for byte.

#2. The conflict grouping was the fourth consumer of the winner rule and
the one that did not take it: an incomplete declaration grouped as an
empty default, so a broken declaration beside a valid one reported a
conflict between two skills where only one had said anything, on a key
that seeded correctly. `writable_all` is now the single gate on
`complete()` and `writable_for` is its first element; merge, record_seeds,
template_for and conflict_notes all take declarations only through it.
There is no way left to obtain one without the rule.

#3. The refusal named "a string or array" while also firing for an
unclosed inline table, sending somebody to the wrong part of their
template. It now says the template opens something never closed, naming
no delimiter — completeness comes off the enumerated grammar, so any list
here goes stale the next time that grammar grows. The key name stays.

Controls, each red against the code it replaces: an LF template seeded
into a CRLF file and the reverse, asserting the value's bytes survive
while the lines around it take the file's EOL; a broken declaration
before a valid one asserting no note at all; and a refusal that names the
key and no delimiter, checked across all four open forms.

settings_seed.rs and its tests were over the size ratchet again, so the
writer moved to write.rs beside env/notes/refresh and the refresh tests
to tests/refresh.rs. Third round of shaving prose to fit, so this splits
on the module's seams instead.
